feat: agregar scripts de seed para usuarios y productos en Prisma
