+++ bknd-nest-11/prisma/scripts/seed.ts
@@ -1,0 +1,21 @@
+import { PrismaClient } from '../generated/prisma/client';
+import { seedProducts } from '../seeds/products';
+import { seedUsers } from '../seeds/users';
+
+const prisma = new PrismaClient();
+
+async function main() {
+  await seedUsers(prisma);
+  await seedProducts(prisma);
+  // Agrega aquí otros seeds si los tienes
+}
+
+main()
+  .then(async () => { await prisma.$disconnect(); })
+  .catch((e) => {
+    console.error(e);
+    process.exit(1);
+  })
+  .finally(async () => {
+    await prisma.$disconnect();
+  });
